==> dev/04.1_Annotate_gene-symbol_sample-id.ipynb <==
# Translate the Ensembl Gene IDs of the data object into Gene Symbols and add as key, var, or obs

## Load adata after processing by 01 to 04

import scanpy as sc
import matplotlib.pyplot as plt
import anndata as ad


import pandas as pd

#adata = sc.read("/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/clustered_adata.h5ad")
#adata = sc.read("/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/filtered_preprocessed_combined_adata.h5ad")
#adata = sc.read("/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/hvg_adata.h5ad")
adata = sc.read("/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/cellcycle_adata.h5ad")
#adata = sc.read("/storage/users/sac43cg/res_Samantha_1/outs/h5as_files/adata_neu.h5ad")  # Replace with your data file

adata

#### Create dictionary

from biomart import BiomartServer

# Connect to the BioMart server
server = BiomartServer("http://ensembl.org/biomart")

# Select the human genes dataset
dataset = server.datasets['hsapiens_gene_ensembl']

# Query BioMart to get the translation from Ensembl ID to gene symbol
response = dataset.search({
    'attributes': ['ensembl_gene_id', 'external_gene_name'],
})

# Initialize an empty dictionary to store the mapping
ensembl_to_gene_name = {}

# Iterate through the response lines
for line in response.iter_lines():
    # Decode the line and split it into Ensembl ID and gene name
    decoded_line = line.decode('utf-8')
    ensembl_id, gene_name = decoded_line.split('\t')
    
    # Store the mapping in the dictionary
    ensembl_to_gene_name[ensembl_id] = gene_name

# Now, ensembl_to_gene_name is a dictionary with Ensembl IDs as keys and gene names as values.


## Name Var_names or var->gene.symbol

adata.var['gene_symbol'] = [ensembl_to_gene_name.get(gene_id, 'NA') for gene_id in adata.var_names]

adata.var_names = [ensembl_to_gene_name.get(gene_id, gene_id) for gene_id in adata.var_names]

adata

# optional
adata.var['gene_symbol']

adata.var_names

## Names Sample IDs as obs

adata.obs_names

### Build a new Index by combining Barcode and Conditions

#df = pd.DataFrame(adata.X, index=adata.obs.index, columns=adata.var.index)

#df.index = df.index + '_' + adata.obs['conditions'].values


def short_hash(input_string, length=8):
    """Get a shorter hash representation of the input string."""
    return str(hash(input_string))[:length]

adata.obs['index'] = [short_hash(name) for name in adata.obs_names]

adata.obs['IDs'] = [f"{condition}_{name}" for condition, name in zip(adata.obs['condition'], adata.obs['index'] )]

#optional
adata.obs_names = adata.obs['IDs']

adata.obs['IDs']

adata.obs_names

adata.var_names

## Export

#adata.write('/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/annotated_adata.h5ad')
adata.write('/storage/users/sac43cg/res_Samantha_1/outs/per_sample_outs/annotated_cellcycle_adata.h5ad')

# Snippets

## If you have a dictionary of Ensembl you can extratc the counts matrix and annotate 
## the columns and rows with a Sample ID and Gene Symbols

# 1. Extracting data
df = pd.DataFrame(adata.X, index=adata.obs.index, columns=adata.var.index)

# 2. Combining barcodes and conditions
df.index = df.index + '_' + adata.obs['conditions'].values

# 3. Translate Ensembl IDs to Gene Symbols
# Assuming ensembl_to_symbol is your dictionary of Ensembl ID -> Gene Symbol
df.columns = df.columns.map(lambda x: ensembl_to_symbol.get(x, x))

counts = pd.DataFrame(adata.X, index=adata.obs_names, columns=adata.var_names)
symbol_adata=sc.AnnData(counts)
